requirements.txt
asgiref==3.11.0
Django==6.0
django-unfold==0.74.1
pillow==12.0.0
sqlparse==0.5.5
